feat(app/operator/sign-up/add-stack): improve input masks
