+++ packages/apps/human-app/frontend/src/i18n/en.json
@@ -2,7 +2,7 @@
   "humanCurrencySymbol": "HMT",
   "inputMasks": {
     "humanCurrencySuffix": " HMT",
-    "percentSuffix": " %",
+    "percentSuffix": "%",
     "plusPrefix": "+"
   },
   "dashboard": {},
